Display Forex data
1) Add a analysis hierachies
2) Add a Forex display PP which carryies out a query for each cell and returns the corresponding rate.

Still many bugs to be fixed :

1) In the definition of ForexHier :
We choose to not set allMember to false. Indeed it does not make any
sense to have permanently a filter with a currency. Because 1) this
filter has no chance to work given that the forex datastore could not be
reached. Because 2) if it worked it would supersede the context value.
But with this choice one could have a "Grand Total" column. Or it's
meaningless to add up currency ! This problem must be fix by introducing
a post-processor.

2) In the Forex Diplay PP :
2.1) When we are looking for the currency of the current cell, we
use a very ugly way : we use the map and then an iterator. And once we
have found the coordinate of the hierarchy, we must retrieve 1 to match
with the level set (Measure does not count anymore)
Is there a faster method ?
2.2) To find the wished forex level coordinate, we must compare the
hierarchy name with a hard coded string ("ForexHier"). And what if one
changes the hierarchy name ?
They could be two way to fix it :
- either we use a global variable
- either we add a key parameter to the PP
2.3) When we display the forex data, it does not make any sense to
display through other dimension than ForexDim (and maybe Time ? depends
of the data file).
Thus we need to forbid the view of the other dimension in ActiveUI when
display the rate.
2.4) We are always considering the reference currency is EUR. But what
if one add a line like RUB->EUR : x.x in the Forex datastore ? In
activeUI, one will see EUR : x.x => no sense ! Should we consider that
EUR will always be the reference currency ?
To my mind, we should better gain in modularity.

3) Same problem than 2.4) but for computing the forex values. What if we
have in the Forex store a line like RUB->EUR : x.x and not the contrary
?

4) When one compute the forex value, what if a record is null ? If there
is no data ?
Should we better return null or the default value ? But is it a true
issue ? To think ...