Forex TuplePublisher to correct Forex data format
The Forex Tuple Publisher corrects the Forex data before publishing them in the store.
It consider the first currency (noted REF) of the first line as the reference currency.
Thus, every following rate will be expressed as : REF -> CUR : ...
The currency can be expressed in whatever order, with whatever links between currency with the
condition that they form a single currency tree.

Warning :
1)   	If the user changes the reference currency during the cube life time this will cause an error !
	Thus, if one wish to change the reference currency one will have to recompile the cube and
	change every place of the project where the reference currency was hard-coded.
	That is to say in where the PP are called i.e in Measures.xml. The query are carryied out
	by supposing that the reference currency is that one.
	If one has rebuild the ForexStore with an other reference currency, the query will look for
	a record having as a key the depracated reference currency.
2)   	The currency relations without link with the reference currency will be ignored
3)	The Tuple Publisher does not manage the error case of a multiple currency definition