Upgrading to 5.6.5 : step 5
Add role context config : definition of the admin role
By this way it allows to define a (default) currency context value.

Moreover, we splitted the construction of the CurrencyContextValueTranslator :
first the constructor, secondly the componenents initialization.
Indeed, the constructor is called before the DatastoreBean and data loading.
Thus the initialisation of the Translator must be called after them !

Results :
- we have a role context config
- the CurrencyContextValueTranslator is working